+++ include/nstd/gl/def.h
@@ -2,6 +2,7 @@
 #define NSTD_GL_DEF_H_INCLUDED
 #include "../core/def.h"
 #include "../nstd.h"
+#include "../string.h"
 
 /// Represents a color.
 // Must match https://docs.rs/wgpu/0.12.0/wgpu/struct.Color.html.
@@ -58,5 +59,35 @@ typedef enum
     /// Use high GPU power.
     NSTD_GL_POWER_PREFERENCE_HIGH
 } NSTDGLPowerPreference;
+
+/// Represents a device type.
+typedef enum
+{
+    /// An unknown device type.
+    NSTD_GL_DEVICE_TYPE_UNKNOWN,
+    /// `wgpu`'s integrated GPU.
+    NSTD_GL_DEVICE_TYPE_INTEGRATED_GPU,
+    /// A physical GPU.
+    NSTD_GL_DEVICE_TYPE_DISCRETE_GPU,
+    /// A virtual/hosted GPU.
+    NSTD_GL_DEVICE_TYPE_VIRTUAL_GPU,
+    /// CPU/Software rendering.
+    NSTD_GL_DEVICE_TYPE_CPU
+} NSTDGLDeviceType;
+
+/// Contains information on a device.
+typedef struct
+{
+    /// The name of the drawing device.
+    NSTDString name;
+    /// The device's vendor.
+    NSTDUSize vendor;
+    /// The ID of the device adapter.
+    NSTDUSize device;
+    /// The type of drawing device.
+    NSTDGLDeviceType device_type;
+    /// The drawing backend in use.
+    NSTDGLBackend backend;
+} NSTDGLDeviceInfo;
 
 #endif