+++ include/nstd/gl/device/device.h
@@ -1,0 +1,35 @@
+#ifndef NSTD_GL_DEVICE_DEVICE_H_INCLUDED
+#define NSTD_GL_DEVICE_DEVICE_H_INCLUDED
+#include "../../core/def.h"
+#include "../../nstd.h"
+#include "../def.h"
+#include "handle.h"
+NSTDCPPSTART
+
+/// Represents a graphics device.
+typedef struct
+{
+    /// A raw pointer to the `wgpu` device.
+    NSTDAny raw;
+    /// The device's command queue.
+    NSTDAny command_queue;
+} NSTDGLDevice;
+
+/// Creates a new device.
+/// Parameters:
+///     `const NSTDGLDeviceHandle device_handle` - A handle to the device.
+/// Returns: `NSTDGLDevice device` - The physical device.
+NSTDAPI NSTDGLDevice nstd_gl_device_new(const NSTDGLDeviceHandle device_handle);
+
+/// Frees a device.
+/// Parameters:
+///     `NSTDGLDevice *const device` - A pointer to the device to free.
+NSTDAPI void nstd_gl_device_free(NSTDGLDevice *const device);
+
+/// Frees an `NSTDGLDeviceInfo` object.
+/// Parameters:
+///     `NSTDGLDeviceInfo *const device_info` - Pointer to an `NSTDGLDeviceInfo` object.
+NSTDAPI void nstd_gl_device_info_free(NSTDGLDeviceInfo *const device_info);
+
+NSTDCPPEND
+#endif